+++ include/irr/video/video.h
@@ -5,9 +5,6 @@
 #include "irr/asset/asset.h"
 #include "irr/ui/ui.h" // unsure yet
 
-//texture packer
-#include "irr/video/IGPUTexturePacker.h"
-
 // alloc
 #include "irr/video/alloc/GPUMemoryAllocatorBase.h"
 #include "irr/video/alloc/HostDeviceMirrorBufferAllocator.h"
@@ -29,5 +26,8 @@
 
 // kill/refactor
 #include "irr/video/CGPUSkinnedMesh.h"
+
+//VT
+#include "irr/video/IGPUVirtualTexture.h"
 
 #endif
